test: split world-buy-client tests by concern
